+++ static/js/global.js
@@ -15,10 +15,4 @@ $(document).ready(function(){
             return false;
         }
     });
-    MathJax.Hub.Config({
-        tex2jax: {
-            inlineMath: [['$','$'], ['\\(','\\)']],
-            processEscapes: true,
-        },
-    });
 });